+++ src/Exceptions/FileExistsException.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Nevadskiy\Downloader\Exceptions;
+
+use Exception;
+
+class FileExistsException extends Exception
+{
+    /**
+     * A path of the file.
+     *
+     * @var string
+     */
+    protected $path;
+
+    /**
+     * Make a new exception instance.
+     */
+    public function __construct(string $path)
+    {
+        $this->path = $path;
+
+        parent::__construct(sprintf('File "%s" already exists', $this->getPath()));
+    }
+
+    /**
+     * Get a path of the file.
+     */
+    public function getPath(): string
+    {
+        return $this->path;
+    }
+}