+++ src/Exceptions/NetworkException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Nevadskiy\Downloader\Exceptions;
+
+use RuntimeException;
+
+class NetworkException extends RuntimeException
+{
+
+}