+++ tests/TestCase.php
@@ -5,6 +5,7 @@ namespace Nevadskiy\Downloader\Tests;
 use Nevadskiy\Downloader\Tests\Constraint\DirectoryIsEmpty;
 use Nevadskiy\Downloader\Tests\Uses\TestingDirectory;
 use PHPUnit\Framework\TestCase as BaseTestCase;
+use const DIRECTORY_SEPARATOR;
 
 class TestCase extends BaseTestCase
 {
@@ -13,9 +14,11 @@ class TestCase extends BaseTestCase
     /**
      * Get a base URL of the server with fixture files.
      */
-    protected function serverUrl(): string
+    protected function serverUrl(string $path = '/'): string
     {
-        return $_ENV['TESTING_SERVER_URL'] ?? 'http://127.0.0.1:8888';
+        $url = $_ENV['TESTING_SERVER_URL'] ?? 'http://127.0.0.1:8888';
+
+        return $url . DIRECTORY_SEPARATOR . ltrim($path, '/');
     }
 
     /**